backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT[crypto]
requests
supabase
